backend/requirements.txt
# Backend Dependencies - ForensicVideo AI Platform

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.4

# Redis & Celery
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.1.0

# AI & Computer Vision
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0

# YOLO (Ultralytics)
ultralytics==8.1.11

# DeepFace
deepface==0.0.87
tf-keras==2.16.0
tensorflow==2.16.1
retina-face==0.0.13

# Real-ESRGAN
realesrgan==0.3.0
basicsr==1.4.2

# Image Processing
scikit-image==0.22.0
imageio==2.33.1

# AWS S3
boto3==1.34.34
botocore==1.34.34

# Google Cloud Storage (optional)
google-cloud-storage==2.14.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
aiofiles==23.2.1

# Monitoring & Logging
sentry-sdk==1.40.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
